Initial upload
This is a multi-threaded, multi-tabbed, SFTP-only client.  I started adding some SSH console functionality to it but that isn't finished at this time.  I used ChatGPT to design each individual function in the application but then sewed it together myself.  No actual design was done, everything was coded and then re-coded on the fly.  The code utilizes a very top-down object-oriented design and is my first application in Python to do so.  A lot of 'shepherding' was done with ChatGPT to try to make the code look and function consistently throughout the source.  
The program is also multi threaded and has a separate thread for downloading files, listing directories, creating directories, etc.  It does not maintain any persistent connections but instead keep track in memory of where in the remote folder structure we're browsing at using internal tables for each open tab.
It also is a multi-window application utilizing QT5 and with various graphical elements like pull-down lists, a site manager, json configuration files, graphical status bars, etc.  I would probably consider this my most mature and 'proper' Python application.